ocr-service/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python
numpy
paddlepaddle
paddleocr
groq
python-dotenv
httpx
